Use local Horizon in integration tests
