+++ .circleci/config.yml
@@ -197,14 +197,6 @@ commands:
           name: Build release artifacts
           command: go run ./support/scripts/build_release_artifacts/main.go
 
-  # build_horizon builds horizon binary
-  build_horizon:
-    steps:
-      - run:
-          name: Build horizon binary
-          command: |
-            go build -v ./services/horizon
-
   # send_coverage_report sends coverage report to codecov.io
   send_coverage_report:
     steps:
@@ -219,6 +211,10 @@ commands:
         type: boolean
         default: false
     steps:
+      - unless:
+          condition: << parameters.enable-captive-core >>
+          steps:
+            - run: docker run -d --env POSTGRES_HOST_AUTH_METHOD=trust -p 5432:5432 circleci/postgres:9.6.5-alpine
       - run:
           name: Run Horizon integration tests <<# parameters.enable-captive-core >>(With captive core)<</ parameters.enable-captive-core >>
           # Currently all integration tests are in a single directory.
