+++ services/horizon/cmd/serve.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
+	horizon "github.com/stellar/go/services/horizon/internal"
 )
 
 var serveCmd = &cobra.Command{
@@ -9,7 +10,7 @@ var serveCmd = &cobra.Command{
 	Short: "run horizon server",
 	Long:  "serve initializes then starts the horizon HTTP server",
 	Run: func(cmd *cobra.Command, args []string) {
-		initApp().Serve()
+		horizon.NewAppFromFlags(config, flags).Serve()
 	},
 }
